requirements.txt
pandas==2.2.3
pillow==11.0.0
setuptools==75.8.0
opencv-python==4.11.0.86
scikit-learn==1.6.1
scipy==1.15.2
pickleshare==0.7.5
numpy==1.26.4         
keras==3.8.0
torch==2.6.0
pyaml==25.1.0
matplotlib==3.10.0
tqdm==4.67.0
pyarrow==19.0.0
torchvision==0.21.0 
timm==0.9.16